+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Virtual list
 - `Referendums list`, `External list`, `Proposals list` on `Democracy page`
 - Progress info block
+- `Proposals list` on `Treasury overview tab`
 
 #### Fixed
 - Collected winning data on auction tab